Add results module
A 'result' is a simple DTO object that wraps certain function calls
that need determination of success/failure with additional metadata
such as stdout/stderr.

For example, imagine we're checking if a repository is dirty. The caller
will expect both an indication if the repo is clean/dirty and, if dirty,
some additional data about which files need to be examined. For cases like
this, we would want to return a Result.