Fix file ignore during graft
This would previously work when the layout directory was a relative
path, e.g. "." and fail when it was a fully qualified path. Now, when
processing directories during copy, we use relative subdir path that
we can join to the src/layout directories to correctly check if those
files exist.